Fixes list view render bug if month has no entries
Closes #4

The logic for splitting the current month into blocks was running
regardless of whether the month had entries or not. The check against
the current month was _within_ the split function, rather than outside of
it. This meant that `null` was being returned and passed through the chain
of subsequent functions.